refactor: move namespace closer to folder structure
