+++ HospitalSystem.ConsoleUI/Director/IngredientsUI.cs
@@ -1,6 +1,4 @@
-using HospitalSystem.Core;
-
-namespace HospitalSystem.ConsoleUI;
+namespace HospitalSystem.ConsoleUI.Director;
 
 public class IngredientsUI : ConsoleUI
 {
