+++ HospitalSystem.Core/Renovations/RenovationService.cs
@@ -1,6 +1,6 @@
 using HospitalSystem.Core.Utils;
 
-namespace HospitalSystem.Core;
+namespace HospitalSystem.Core.Renovations;
 
 [System.Serializable]
 public class RenovationException : System.Exception
